+++ sailr_ext.h
@@ -15,5 +15,7 @@ ext_func_hash_object* sailr_ext_func_hash_init();
 void sailr_ext_func_hash_add(ext_func_hash_object** hash, const char* fname, unsigned int num_args, int (* func)(arg_list*, unsigned int, vm_stack*));
 void sailr_ext_func_hash_free(ext_func_hash_object** hash);
 
+const char* sailr_ext_func_hash_get_last_executed(ext_func_hash_object** hash);
+void sailr_ext_func_hash_reset_last_executed(ext_func_hash_object** hash);
 #endif